+++ Model/Util.php
@@ -27,6 +27,9 @@ class Util {
     // Store configuration key for Flow API Token
     const FLOW_API_TOKEN = 'flowcommerce/flowconnector/api_token';
 
+    // Store configuration key for Flow Invoice Event
+    const FLOW_INVOICE_EVENT = 'flowcommerce/flowconnector/invoice_event';
+
     // Flow API base endpoint
     const FLOW_API_BASE_ENDPOINT = 'https://api.flow.io/';
 
@@ -228,9 +231,11 @@ class Util {
 
     /**
      * Returns the ID of the current store
+     * @return int
+     * @throws \Magento\Framework\Exception\NoSuchEntityException
      */
     public function getCurrentStoreId() {
-        $this->storeManager->getStore()->getId();
+        return $this->storeManager->getStore()->getId();
     }
 
     /**
@@ -283,5 +288,21 @@ class Util {
             }
         }
         return $stores;
+    }
+
+    /**
+     * Returns Flow Invoice Event
+     *
+     * @see \FlowCommerce\FlowConnector\Model\Config\Source\InvoiceEvent::toOptionArray()
+     *
+     * @param null $storeId
+     * @return int
+     * @throws \Magento\Framework\Exception\NoSuchEntityException
+     */
+    public function getFlowInvoiceEvent($storeId = null) {
+        if ($storeId === null) {
+            $storeId = $this->getCurrentStoreId();
+        }
+        return (int)$this->scopeConfig->getValue(self::FLOW_INVOICE_EVENT, ScopeInterface::SCOPE_STORE, $storeId);
     }
 }