+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Flow Connector for Magento 2 Change Log
 
+## 1.1.0
+
+- Magento invoices and shipments can now be created automatically.
+- Magento now handles storing discounts better for all sales entities.
+
 ## 1.0.40
 
 - Fix loading customers by email on orders